Feature request 5852: Option to toggle items in shutdown menu (1/2)
This allows for the selective switching of the various CM additions on the
power menu.  Reboot and Profiles are enabled by default, Screenshot is
disabled.

This is the completion of the feature that was merged while still partially
complete.

Patchset 1-2: Initial commit
Patchset 3  : Fix System settings npe
Patchset 4  : Cleanup imports

Change-Id: I6ebcbf684da5724f34d65180d9e808e953553a42